requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
Pygments==2.17.2
sqlparse==0.4.4
typing_extensions==4.10.0
